Clean up warnings
